done with cleanup

+++ src/main/java/io/turntabl/model/metrics/GCDuration.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 public class GCDuration extends GarbageCollection {
 
-    public GCDuration(String name, String type, HashMap<String, Double> value, Long timestamp, int interval, HashMap<String, String> attributes) {
+    public GCDuration(String name, String type, HashMap<String, Double> value, Long timestamp, Long interval, HashMap<String, String> attributes) {
         super(name, type, value, timestamp, interval, attributes);
     }
