+++ src/main/java/io/turntabl/model/metrics/GCMajorDuration.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 public class GCMajorDuration extends GarbageCollection {
 
-    public GCMajorDuration(String name, String type, HashMap<String, Double> value, Long timestamp, int interval, HashMap<String, String> attributes) {
+    public GCMajorDuration(String name, String type, HashMap<String, Double> value, Long timestamp, Long interval, HashMap<String, String> attributes) {
         super(name, type, value, timestamp, interval, attributes);
     }
